Fixes an unwanted error "user already exists".
Doing a query with an AND SQL statement needs to be done with multiple "filter" callswith SQLAlchemy.

Here, we want to be sure that the username is not used AND that the project is the same than the eventual users that would match. The previous version of the code returned an user with the same name, even if the user wasn't in the right group.