Add .PHONY instructions to the Makefile
`.PHONY` instructs Make to not look, for instance, for an `install` file
before executing the corresponding target. If such a file would exist,
the command would not be performed at all. This is because Make is
initially intended to create files (i.e. targets) and consider there is
nothing to do if the file already exists.

More information on https://stackoverflow.com/a/2145605.